+++ src/main/java/com/smartsheet/api/models/LabelApproverEntry.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright (C) 2025 Smartsheet
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.smartsheet.api.models;
+
+import java.util.List;
+
+/**
+ * Represents a per-label approver entry in downgrade approval settings.
+ */
+public class LabelApproverEntry {
+
+    private String labelId;
+    private List<ApproverEntry> approvers;
+
+    public String getLabelId() {
+        return labelId;
+    }
+
+    public void setLabelId(String labelId) {
+        this.labelId = labelId;
+    }
+
+    public List<ApproverEntry> getApprovers() {
+        return approvers;
+    }
+
+    public void setApprovers(List<ApproverEntry> approvers) {
+        this.approvers = approvers;
+    }
+}