+++ src/test/java/com/smartsheet/api/sdktest/governance/CommonTestConstants.java
@@ -1,0 +1,23 @@
+/*
+ * Copyright (C) 2025 Smartsheet
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.smartsheet.api.sdktest.governance;
+
+public class CommonTestConstants {
+    public static final long TEST_PLAN_ID = 1148023251199876L;
+    public static final long TEST_ORG_ID = 1556806293055364L;
+    public static final long TEST_ASSET_ID = 112398785741L;
+}